seperate into distinct libraries
The myt and abrp api might be useful
beyond my use and outside of spin.
Start with the myt api work to put outside.
