+++ Cargo.toml
@@ -1,26 +1,6 @@
-[package]
-name = "myt2abrp"
-authors = ["Ralf Anton Beier <ralf_beier@me.com>"]
-description = "Connector between Toyoto MyT (Europe) and ABRP"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = [ "cdylib" ]
-
-[dependencies]
-# Useful crate to handle errors.
-anyhow = "1"
-# Crate to simplify working with bytes.
-bytes = "1"
-# General-purpose crate with common HTTP types.
-http = "0.2"
-# The Spin SDK.
-spin-sdk = { git = "https://github.com/fermyon/spin", tag = "v1.0.0" }
-# Crate that generates Rust Wasm bindings from a WebAssembly interface.
-wit-bindgen-rust = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "cb871cfa1ee460b51eb1d144b175b9aab9c50aba" }
-
-serde = { version = "1.0.25", features = ["derive"]}
-serde_json = { version = "1.0" }
-tracing = { version = "0.1.25", features = ["log"] }
 [workspace]
+
+members = [
+    "myt2abrp",
+    "myt",
+]
